+++ tanjun/parsing.py
@@ -56,6 +56,8 @@ import shlex
 import typing
 from collections import abc as collections
 
+import typing_extensions
+
 from . import abc as tanjun
 from . import conversion
 from . import errors
@@ -86,8 +88,6 @@ _T = typing.TypeVar("_T")
 
 # 3.9 and 3.10 just can't handle ending Concatenate with ... so we lie about this at runtime.
 if typing.TYPE_CHECKING:
-    import typing_extensions
-
     _P = typing_extensions.ParamSpec("_P")
     _ConverterSig = collections.Callable[
         typing_extensions.Concatenate[str, _P], typing.Union[collections.Coroutine[typing.Any, typing.Any, _T], _T]
@@ -129,7 +129,7 @@ class UndefinedT:
 
 
 UndefinedDefaultT = UndefinedT
-"""Deprecated alias of [Undefined][tanjun.parsing.UNDEFINED]."""
+"""Deprecated alias of [UndefinedT][tanjun.parsing.UndefinedT]."""
 
 UNDEFINED = UndefinedT()
 """A singleton used to represent an undefined value within parsing logic."""
@@ -475,8 +475,11 @@ class AbstractOptionParser(tanjun.MessageParser, abc.ABC):
         """
 
 
-AbstractParser = AbstractOptionParser
-"""Deprecated alias of [AbstractOptionParser][tanjun.parsing.AbstractOptionParser]."""
+@typing_extensions.deprecated("Use AbstractOptionParser")
+class AbstractParser(AbstractOptionParser):
+    """Deprecated alias of [AbstractOptionParser][tanjun.parsing.AbstractOptionParser]."""
+
+    __slots__ = ()
 
 
 class _ShlexTokenizer: