+++ tests/test_conversion.py
@@ -52,10 +52,12 @@ class TestBaseConverter:
         ...
 
     def test_async_caches_property(self):
-        assert tanjun.conversion.BaseConverter().async_caches == []
+        with pytest.warns(DeprecationWarning, match="Use .caches instead"):
+            assert tanjun.conversion.BaseConverter().async_caches == []
 
     def test_cache_components_property(self):
-        assert tanjun.conversion.BaseConverter().cache_components is hikari.api.CacheComponents.NONE
+        with pytest.warns(DeprecationWarning, match="Use .caches instead"):
+            assert tanjun.conversion.BaseConverter().cache_components is hikari.api.CacheComponents.NONE
 
     @pytest.mark.parametrize(
         ("obj", "expected"),
@@ -175,7 +177,8 @@ class TestBaseConverter:
         assert obj.caches == expected
 
     def test_intents_property(self):
-        assert tanjun.conversion.BaseConverter().intents is hikari.Intents.NONE
+        with pytest.warns(DeprecationWarning, match="Use .caches instead"):
+            assert tanjun.conversion.BaseConverter().intents is hikari.Intents.NONE
 
     @pytest.mark.parametrize(
         ("obj", "expected"),
@@ -567,7 +570,7 @@ class TestToChannel:
         mock_dm_cache = mock.AsyncMock()
         mock_thread_cache = mock.AsyncMock()
         mock_thread_cache.get.side_effect = tanjun.dependencies.EntryNotFound
-        converter = tanjun.conversion.ChannelConverter(include_dms=False)
+        converter = tanjun.conversion.ToChannel(include_dms=False)
 
         with pytest.raises(ValueError, match="Couldn't find channel"):
             await converter(
@@ -652,7 +655,7 @@ class TestToChannel:
         mock_dm_cache = mock.AsyncMock()
         mock_thread_cache = mock.AsyncMock()
         mock_thread_cache.get.side_effect = tanjun.dependencies.EntryNotFound
-        converter = tanjun.conversion.ChannelConverter(include_dms=False)
+        converter = tanjun.conversion.ToChannel(include_dms=False)
 
         with pytest.raises(ValueError, match="Couldn't find channel"):
             await converter(
@@ -729,7 +732,7 @@ class TestToChannel:
         mock_dm_cache = mock.AsyncMock()
         mock_thread_cache = mock.AsyncMock()
         mock_thread_cache.get.side_effect = tanjun.dependencies.CacheMissError
-        converter = tanjun.conversion.ChannelConverter(include_dms=False)
+        converter = tanjun.conversion.ToChannel(include_dms=False)
 
         result = await converter(
             "<#12222>", mock_context, cache=mock_channel_cache, dm_cache=mock_dm_cache, thread_cache=mock_thread_cache
@@ -752,7 +755,7 @@ class TestToChannel:
         mock_dm_cache = mock.AsyncMock()
         mock_thread_cache = mock.AsyncMock()
         mock_thread_cache.get.side_effect = tanjun.dependencies.CacheMissError
-        converter = tanjun.conversion.ChannelConverter(include_dms=False)
+        converter = tanjun.conversion.ToChannel(include_dms=False)
 
         with pytest.raises(ValueError, match="Only the following channel types are allowed for this argument: .*"):
             await converter(
@@ -780,7 +783,7 @@ class TestToChannel:
         mock_dm_cache.get.side_effect = tanjun.dependencies.CacheMissError
         mock_thread_cache = mock.AsyncMock()
         mock_thread_cache.get.side_effect = tanjun.dependencies.CacheMissError
-        converter = tanjun.conversion.ChannelConverter(include_dms=False)
+        converter = tanjun.conversion.ToChannel(include_dms=False)
 
         with pytest.raises(ValueError, match="Couldn't find channel"):
             await converter(