+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Marked deprecated classes, functions and overloads as deprecated using `typing.deprecated`.
+  Deprecated class aliases are now separate classes to allow warning on init.
+
 ## [2.11.4] - 2023-02-08
 ### Fixed
 - `Context.create_initial_response` (and by extension `Context.respond` for the initial